test: merge test
