+++ application/server/src/test/kotlin/io/github/siyual_park/application/server/controller/PingControllerTest.kt
@@ -2,7 +2,7 @@ package io.github.siyual_park.application.server.controller
 
 import io.github.siyual_park.IntegrationTest
 import io.github.siyual_park.application.server.gateway.PingControllerGateway
-import io.github.siyual_park.coroutine.test.CoroutineTest
+import io.github.siyual_park.coroutine.test.CoroutineTestHelper
 import kotlinx.coroutines.reactive.awaitSingle
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus
 @IntegrationTest
 class PingControllerTest @Autowired constructor(
     private val pingControllerGateway: PingControllerGateway,
-) : CoroutineTest() {
+) : CoroutineTestHelper() {
     @Test
     fun `GET ping, status = 201`() = blocking {
         val response = pingControllerGateway.ping()
