+++ domain/auth/build.gradle.kts
@@ -5,8 +5,6 @@ dependencies {
     implementation(project(":module:persistence"))
     implementation(project(":module:mapper"))
     implementation(project(":module:util"))
-
-    testImplementation(project(":module:data-test"))
 }
 
 kotlin.sourceSets["main"].kotlin.srcDirs("src/main")
