--- module/data-test/.gitignore
@@ -1,7 +1,0 @@
-/.gradle
-/.idea
-/out
-/build
-*.iml
-*.ipr
-*.iws
