--- module/data-test/src/test/kotlin/io/github/siyual_park/data/test/entity/Person.kt
@@ -1,14 +1,0 @@
-package io.github.siyual_park.data.test.entity
-
-import io.github.siyual_park.data.ModifiableULIDEntity
-import io.github.siyual_park.data.annotation.Key
-import org.springframework.data.mongodb.core.mapping.Document
-import org.springframework.data.relational.core.mapping.Table
-
-@Table("persons")
-@Document("persons")
-data class Person(
-    @Key
-    var name: String,
-    var age: Int
-) : ModifiableULIDEntity()
