+++ module/data/src/main/kotlin/io/github/siyual_park/data/test/DataTestHelper.kt
@@ -1,0 +1,106 @@
+package io.github.siyual_park.data.test
+
+import com.mongodb.reactivestreams.client.MongoClient
+import de.flapdoodle.embed.mongo.MongodExecutable
+import io.github.siyual_park.coroutine.test.CoroutineTestHelper
+import io.github.siyual_park.data.converter.BinaryToULIDConverter
+import io.github.siyual_park.data.converter.BytesToULIDConverter
+import io.github.siyual_park.data.converter.ULIDToBinaryConverter
+import io.github.siyual_park.data.converter.ULIDToBytesConverter
+import io.github.siyual_park.data.migration.MigrationManager
+import io.r2dbc.h2.H2ConnectionFactory
+import kotlinx.coroutines.CoroutineScope
+import org.junit.jupiter.api.AfterAll
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.springframework.core.convert.converter.Converter
+import org.springframework.data.mongodb.core.ReactiveMongoTemplate
+import org.springframework.r2dbc.connection.R2dbcTransactionManager
+import org.springframework.transaction.ReactiveTransaction
+import org.springframework.transaction.reactive.TransactionalOperator
+import org.springframework.transaction.reactive.executeAndAwait
+import java.util.UUID
+
+abstract class DataTestHelper(
+    converters: Collection<Converter<*, *>> = emptyList()
+) : CoroutineTestHelper() {
+    private val database: String = UUID.randomUUID().toString()
+
+    private val connectionFactory = H2ConnectionFactory.inMemory(database)
+    val entityOperations = createR2dbcEntityTemplate(
+        connectionFactory,
+        mutableListOf<Converter<*, *>>(
+            ULIDToBytesConverter(),
+            BytesToULIDConverter(),
+        ).also {
+            it.addAll(converters)
+        }
+    )
+
+    val migrationManager = MigrationManager(entityOperations)
+
+    var transactionManager = R2dbcTransactionManager(connectionFactory)
+    var transactionalOperator = TransactionalOperator.create(transactionManager)
+
+    @BeforeEach
+    override fun setUp() {
+        super.setUp()
+
+        blocking {
+            migrationManager.run()
+        }
+    }
+
+    @AfterEach
+    override fun tearDown() {
+        blocking {
+            migrationManager.revert()
+        }
+
+        super.tearDown()
+    }
+
+    @Test
+    fun contextLoads() {
+    }
+
+    fun transactional(func: suspend CoroutineScope.(ReactiveTransaction) -> Unit) = blocking {
+        transactionalOperator.executeAndAwait {
+            func(it)
+        }
+    }
+
+    companion object {
+        lateinit var mongodExecutable: MongodExecutable
+        lateinit var mongoClient: MongoClient
+        lateinit var mongoTemplate: ReactiveMongoTemplate
+
+        private val database = UUID.randomUUID().toString()
+
+        @BeforeAll
+        @JvmStatic
+        fun setupMongo() {
+            val mongodExecutableAndClient = createEmbeddedMongoDBClients()
+            mongodExecutable = mongodExecutableAndClient.first
+            mongodExecutable.start()
+
+            mongoClient = mongodExecutableAndClient.second
+            mongoTemplate = createReactiveMongoTemplate(
+                mongoClient,
+                database,
+                mutableListOf<Converter<*, *>>(
+                    BinaryToULIDConverter(),
+                    ULIDToBinaryConverter(),
+                )
+            )
+        }
+
+        @AfterAll
+        @JvmStatic
+        fun stopMongo() {
+            mongodExecutable.stop()
+        }
+    }
+}
