+++ module/data/src/main/kotlin/io/github/siyual_park/data/test/R2dbcEntityTemplate.kt
