+++ module/data/src/test/kotlin/io/github/siyual_park/data/repository/r2dbc/entity/Person.kt
@@ -1,0 +1,16 @@
+package io.github.siyual_park.data.repository.r2dbc.entity
+
+import io.github.siyual_park.data.ModifiableULIDEntity
+import io.github.siyual_park.data.annotation.Key
+import org.springframework.data.mongodb.core.mapping.Document
+import org.springframework.data.relational.core.mapping.Table
+import java.time.Instant
+
+@Table("persons")
+@Document("persons")
+data class Person(
+    @Key
+    var name: String,
+    var age: Int,
+    var deletedAt: Instant? = null
+) : ModifiableULIDEntity()
