+++ module/data/src/test/kotlin/io/github/siyual_park/data/repository/r2dbc/repository/cache/InMemoryNestedStorageTest.kt
@@ -1,0 +1,143 @@
+package io.github.siyual_park.data.repository.r2dbc.repository.cache
+
+import com.google.common.cache.CacheBuilder
+import io.github.siyual_park.data.repository.cache.Extractor
+import io.github.siyual_park.data.repository.cache.InMemoryNestedStorage
+import io.github.siyual_park.data.repository.r2dbc.SimpleR2DBCRepository
+import io.github.siyual_park.data.repository.r2dbc.dummy.DummyPerson
+import io.github.siyual_park.data.repository.r2dbc.entity.Person
+import io.github.siyual_park.data.repository.r2dbc.repository.r2dbc.migration.CreatePerson
+import io.github.siyual_park.data.test.DataTestHelper
+import io.github.siyual_park.ulid.ULID
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+
+@Suppress("UNCHECKED_CAST")
+class InMemoryNestedStorageTest : DataTestHelper() {
+    private val personRepository = SimpleR2DBCRepository<Person, ULID>(entityOperations, Person::class)
+
+    private val storage = InMemoryNestedStorage(
+        CacheBuilder.newBuilder() as CacheBuilder<ULID, Person>,
+        object : Extractor<Person, ULID> {
+            override fun getKey(entity: Person): ULID {
+                return entity.id
+            }
+        }
+    ).also {
+        it.createIndex(
+            "name",
+            object : Extractor<Person, String> {
+                override fun getKey(entity: Person): String {
+                    return entity.name
+                }
+            }
+        )
+    }
+
+    init {
+        migrationManager.register(CreatePerson(entityOperations))
+    }
+
+    @BeforeEach
+    override fun setUp() {
+        super.setUp()
+
+        storage.clear()
+    }
+
+    @Test
+    fun put() = blocking {
+        val child1 = storage.fork()
+        val child2 = child1.fork()
+
+        val person = DummyPerson.create()
+            .let { personRepository.create(it) }
+
+        child2.put(person)
+
+        assertEquals(person, child2.getIfPresent(person.id))
+        assertEquals(person, child2.getIfPresent("name", person.name))
+
+        assertEquals(null, child1.getIfPresent(person.id))
+        assertEquals(null, child1.getIfPresent("name", person.name))
+
+        assertEquals(null, storage.getIfPresent(person.id))
+        assertEquals(null, storage.getIfPresent("name", person.name))
+
+        child1.merge(child2)
+
+        assertEquals(person, child2.getIfPresent(person.id))
+        assertEquals(person, child2.getIfPresent("name", person.name))
+
+        assertEquals(person, child1.getIfPresent(person.id))
+        assertEquals(person, child1.getIfPresent("name", person.name))
+
+        assertEquals(null, storage.getIfPresent(person.id))
+        assertEquals(null, storage.getIfPresent("name", person.name))
+
+        storage.merge(child1)
+
+        assertEquals(person, child2.getIfPresent(person.id))
+        assertEquals(person, child2.getIfPresent("name", person.name))
+
+        assertEquals(person, child1.getIfPresent(person.id))
+        assertEquals(person, child1.getIfPresent("name", person.name))
+
+        assertEquals(person, storage.getIfPresent(person.id))
+        assertEquals(person, storage.getIfPresent("name", person.name))
+    }
+
+    @Test
+    fun remove() = blocking {
+        val child1 = storage.fork()
+        val child2 = child1.fork()
+
+        val person = DummyPerson.create()
+            .let { personRepository.create(it) }
+
+        storage.put(person)
+
+        assertEquals(person, child2.getIfPresent(person.id))
+        assertEquals(person, child2.getIfPresent("name", person.name))
+
+        assertEquals(person, child1.getIfPresent(person.id))
+        assertEquals(person, child1.getIfPresent("name", person.name))
+
+        assertEquals(person, storage.getIfPresent(person.id))
+        assertEquals(person, storage.getIfPresent("name", person.name))
+
+        child2.remove(person.id)
+
+        assertEquals(null, child2.getIfPresent(person.id))
+        assertEquals(null, child2.getIfPresent("name", person.name))
+
+        assertEquals(person, child1.getIfPresent(person.id))
+        assertEquals(person, child1.getIfPresent("name", person.name))
+
+        assertEquals(person, storage.getIfPresent(person.id))
+        assertEquals(person, storage.getIfPresent("name", person.name))
+
+        child1.merge(child2)
+
+        assertEquals(null, child2.getIfPresent(person.id))
+        assertEquals(null, child2.getIfPresent("name", person.name))
+
+        assertEquals(null, child1.getIfPresent(person.id))
+        assertEquals(null, child1.getIfPresent("name", person.name))
+
+        assertEquals(person, storage.getIfPresent(person.id))
+        assertEquals(person, storage.getIfPresent("name", person.name))
+
+        storage.merge(child1)
+
+        assertEquals(null, child2.getIfPresent(person.id))
+        assertEquals(null, child2.getIfPresent("name", person.name))
+
+        assertEquals(null, child1.getIfPresent(person.id))
+        assertEquals(null, child1.getIfPresent("name", person.name))
+
+        assertEquals(null, storage.getIfPresent(person.id))
+        assertEquals(null, storage.getIfPresent("name", person.name))
+    }
+}
