+++ module/persistence/build.gradle.kts
@@ -4,7 +4,6 @@ dependencies {
 
     implementation(project(":module:ulid"))
     testImplementation(project(":module:coroutine-test"))
-    testImplementation(project(":module:data-test"))
     testImplementation(project(":module:util"))
 }
